visit_type now just calls extract type the function. It's needed because other converters are using the type extraction
